requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
jinja2>=3.1.2
plotly>=5.17.0
tqdm>=4.66.1
pytest>=7.4.2